Agent_kernel/requirements.txt
langgraph
langchain-openai
langchain-core
typing_extensions
ddgs

# 知识库：PDF + Chroma + LlamaIndex（按章节切分，本地 BGE-M3，本地持久化可打包迁移）
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb
unstructured
pdfminer.six
pypdf
